requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
together 